+++ src/containers/NewMember.jsx
@@ -66,8 +66,8 @@ const NewMemberForm = (props) => {
       stairs: '',
       bloc: '',
       postal_code: '',
-      state: undefined,
-      city: undefined,
+      state:{id:'', name: ''},
+      city: {id:'', name: ''},
       lat: undefined,
       long: undefined
     },